Code cleanup set cmake version to 3.9

+++ src/dynamic_string.c
@@ -105,7 +105,7 @@ bool DString_equals(DString str, DString other) {
     if (str->length != other->length) {
         return false;
     }
-    for (int i = 0; i < str->length; ++i) {
+    for (size_t i = 0; i < str->length; ++i) {
         if (str->raw[i] != other->raw[i]) {
             return false;
         }
@@ -117,7 +117,7 @@ bool DString_equalsIgnoreCase(DString str, DString other) {
     if (str->length != other->length) {
         return false;
     }
-    for (int i = 0; i < str->length; ++i) {
+    for (size_t i = 0; i < str->length; ++i) {
         if (UPPER(str->raw[i]) != UPPER(other->raw[i])) {
             return false;
         }
@@ -126,7 +126,7 @@ bool DString_equalsIgnoreCase(DString str, DString other) {
 }
 
 bool DString_rawEquals(DString str, char *other) {
-    int i;
+    size_t i;
     for(i = 0; i < str->length; ++i) {
         if(str->raw[i] != other[i] || other[i] == '\0') {
             return false;
@@ -136,7 +136,7 @@ bool DString_rawEquals(DString str, char *other) {
 }
 
 bool DString_rawEqualsIgnoreCase(DString str, char *other) {
-    int i;
+    size_t i;
     for(i = 0; i < str->length; ++i) {
         if(UPPER(str->raw[i]) != UPPER(other[i]) || other[i] == '\0') {
             return false;
@@ -146,14 +146,14 @@ bool DString_rawEqualsIgnoreCase(DString str, char *other) {
 }
 
 DString DString_toUpperCase(DString str) {
-    for(int i = 0; i < str->length; ++i) {
+    for(size_t i = 0; i < str->length; ++i) {
         str->raw[i] = UPPER(str->raw[i]);
     }
     return str;
 }
 
 int DString_indexOf(DString str, char c) {
-    for(int i = 0; i < str->length; ++i) {
+    for(size_t i = 0; i < str->length; ++i) {
         if(str->raw[i] == c) {
             return i;
         }
